UiTest - Added empty tests to ratings table, misc refactoring and tidy up
- Added two tests to Ratings UI tests with TODOs to be completed
- Renamed UI test classes to remove Fragment terminology
- Added test utils class for the Mockito mock function
- Misc cleanup and linting
