+++ app/src/androidTest/kotlin/com/infinity/foodstandards/TestUtils.kt
@@ -1,0 +1,5 @@
+package com.infinity.foodstandards
+
+import org.mockito.Mockito
+
+internal inline fun <reified T> mock(): T = Mockito.mock(T::class.java)
